editor/requirements.txt
# generated by pip freeze
aiofiles==23.1.0
aiosqlite==0.19.0
async-timeout==4.0.2
asyncua @ git+https://github.com/FreeOpcUa/opcua-asyncio.git@98a64897a2d171653353de2f36d33085aef65e82
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
fonttools==4.39.3
gattrdict==2.0.1
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
incremental==22.10.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
msgpack==1.0.5
Nevow @ git+https://git@github.com/beremiz/nevow-py3.git@cb62cc37824361725c0c0a599802b15210e6aaa3#egg=Nevow
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pycountry==22.3.5
pycparser==2.21
pyparsing==3.0.9
Pyro5==5.14
python-dateutil==2.8.2
pytz==2023.3
serpent==1.41
six==1.16.0
sortedcontainers==2.4.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
wxPython==4.2.1
zeroconf==0.62.0
zope.interface==6.0
